#38 Refactor filtering panel and add refresh button
