+++ src/app.tsx
@@ -5,7 +5,7 @@ import EnterQuizResults from './EnterQuizResults';
 import Footer from './Footer';
 import NavigationBar from './NavigationBar';
 import QuizDetails from './QuizDetails';
-import QuizList from './QuizList';
+import QuizList from './pages/list/QuizList';
 import { useQuizlord } from './QuizlordProvider';
 import Button from './components/Button';
 import Loader from './components/Loader';
