+++ src/hooks/useQuizList.hook.ts
@@ -5,10 +5,11 @@ import { QUIZZES } from '../queries/quiz';
 
 export function useQuizList(isFilteringOnIncomplete: boolean) {
   const { user } = useQuizlord();
-  const { loading, data, fetchMore } = useQuery(QUIZZES, {
+  const { loading, data, fetchMore, refetch } = useQuery(QUIZZES, {
     variables: { filters: isFilteringOnIncomplete ? { excludeCompletedBy: [user?.email] } : {} },
     fetchPolicy: 'network-only',
+    notifyOnNetworkStatusChange: true,
   });
 
-  return { loading, data, fetchMore };
+  return { loading, data, fetchMore, refetch };
 }
