+++ src/pages/list/quizFilters.ts
@@ -1,0 +1,3 @@
+export interface QuizFilters {
+  isFilteringOnIncomplete: boolean;
+}
